Fix error when number of bits is not divisible by 8
When the number of bits in the bitmap is not divisible by 8,
the Bloom filter would allocate one less byte than necessary,
leading to an invalid array access.  Fix this by setting
the number of bytes to the ceiling of the number of bits
divided by 8 instead of the floor.

To be extra safe, also add an assertion that the Bloom filter
parameters are positive numbers.